+++ contracts/interfaces/ILayerZeroUserApplicationConfig.sol
@@ -8,7 +8,12 @@ interface ILayerZeroUserApplicationConfig {
     // @param _chainId - the chainId for the pending config change
     // @param _configType - type of configuration. every messaging library has its own convention.
     // @param _config - configuration in the bytes. can encode arbitrary content.
-    function setConfig(uint16 _version, uint16 _chainId, uint _configType, bytes calldata _config) external;
+    function setConfig(
+        uint16 _version,
+        uint16 _chainId,
+        uint256 _configType,
+        bytes calldata _config
+    ) external;
 
     // @notice set the send() LayerZero messaging library version to _version
     // @param _version - new messaging library version
