+++ contracts/interfaces/ILayerZeroUserApplicationConfigV2.sol
@@ -8,7 +8,12 @@ interface ILayerZeroUserApplicationConfigV2 {
     // @param _chainId - the chainId for the pending config change
     // @param _configType - type of configuration. every messaging library has its own convention.
     // @param _config - configuration in the bytes. can encode arbitrary content.
-    function setConfig(uint16 _version, uint16 _chainId, uint _configType, bytes calldata _config) external;
+    function setConfig(
+        uint16 _version,
+        uint16 _chainId,
+        uint256 _configType,
+        bytes calldata _config
+    ) external;
 
     // @notice set the send() LayerZero messaging library version to _version
     // @param _version - new messaging library version
@@ -28,10 +33,10 @@ interface ILayerZeroUserApplicationConfigV2 {
     // @param _srcAddress - the contract address of the source contract at the source chain
     function setTrustedSource(uint16 _srcChainId, bytes calldata _srcAddress) external;
 
-    // @notice To query if the specified (_srcChainId, _srcAddress) is whitelisted properly. 
+    // @notice To query if the specified (_srcChainId, _srcAddress) is whitelisted properly.
     // @param _srcChainId - the chainId of the source chain
     // @param _srcAddress - the contract address of the source contract at the source chain
-    function isTrustedSource(uint16 _srcChainId, bytes calldata _srcAddress) external view returns(bool);
+    function isTrustedSource(uint16 _srcChainId, bytes calldata _srcAddress) external view returns (bool);
 
     event SetTrustedSource(uint16 _srcChainId, bytes _srcAddress);
 }
